# Use scaling factor to compute ttnn.mean for multi-dim non w,h case + scaling factor fixes  (#28644)