--- a/ttnn/cpp/ttnn/operations/reduction/generic/generic_reductions.cpp
+++ b/ttnn/cpp/ttnn/operations/reduction/generic/generic_reductions.cpp
@@ -244,9 +244,13 @@ static Tensor reduce_impl(
         if (dim.size() == 1 || linear_type) {
             output_tensor = reduce_nd_loop(/*use_reduce_type=*/true);
         } else if constexpr (reduce_type == ReduceType::Mean) {
+            int reduced_volume = 1;
+            for (int axis : dim) {
+                reduced_volume *= input_shape[axis];
+            }
             output_tensor = reduce_nd_loop(
                 /*use_reduce_type=*/false);
-            float inv_volume = 1.0f / input_tensor_arg.logical_volume();
+            float inv_volume = 1.0f / reduced_volume;
             output_tensor = ttnn::mul_sfpu(inv_volume, output_tensor, memory_config);
         } else {
             TT_THROW("Unsupported reduction operation");