@yiding 2025-09-30
Rebased, pulled out the scalar computation, and added a (disabled) test for scalar.