@yiding 2025-09-30
That's one of the tests I added that is averaging ones(), so seed shouldn't matter. FWIW I am using blackhole and test seems like it's on wormhole.