@yiding 2025-10-01
I fixed my bug by putting back the bool first = ... check, since I think it's simpler than trying to intuit whether the dimension is the first dimension in the list of dimensions being reduced.

Also I can re-enable the new test I added for non-unity scaling factor, as the problem I was thinking of only apply to sum and not mean.